Feat/integration (#5)
* import List

* findById(기본 제공 메서드) 삭제

* DB 연동 키 추가.
인텔리제이 화면 위 디버그 오른쪽 점 세개
->
구성 편집
->
옵션 수정
->
환경 변수 설정해서 사용하세요

* 시큐리티 삭제.
롬복 기본 설정 삭제 -> 버그 수정 완료!

* 혹시 몰라서 학번은 다 Long으로 통일했습니다
(Mysql에서 BIGINT로 자동 매핑)

* 로그인 실패 시 RuntimeException 반환

* 사용 안 하는 코드 주석 처리

* User 엔티티
userName -> name

* createReservation: not found 예외 추가

+++ src/main/java/com/konkukrent/demo/dto/LoginResponse.java
@@ -10,6 +10,6 @@ import lombok.*;
 public class LoginResponse {
     private Long userId;
     private String userName;
-    private int studentNum;
+    private Long studentNum;
     private String role;
 }
