+++ src/main/java/com/konkukrent/demo/dto/SignupRequest.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class SignupRequest{
     private String userName;
-    private int studentNum;
+    private Long studentNum;
     private String password;
     private String role;
 }
